fix: Replace EntityActions with EntityAction operators - beta.6
