+++ lib/src/index.ts
@@ -2,7 +2,7 @@
 // NO BARRELS or else `ng build --aot` of any app using ngrx-data produces strange errors
 // actions
 export * from './actions/entity-action';
-export * from './actions/entity-actions';
+export * from './actions/entity-action-operators';
 export * from './actions/entity-action-guard';
 export * from './actions/entity-cache-actions';
 export * from './actions/entity-op';
@@ -57,4 +57,9 @@ export * from './utils/default-logger';
 export * from './utils/default-pluralizer';
 export * from './utils/utilities';
 
-export { NgrxDataModule, NgrxDataModuleConfig } from './ngrx-data.module';
+// NgrxDataModule
+export { NgrxDataModule } from './ngrx-data.module';
+export {
+  NgrxDataModuleWithoutEffects,
+  NgrxDataModuleConfig
+} from './ngrx-data-without-effects.module';
