+++ src/app/store/entity/ngrx-data-toast.service.ts
@@ -1,13 +1,22 @@
 import { Injectable } from '@angular/core';
-import { EntityActions, OP_ERROR, OP_SUCCESS } from 'ngrx-data';
+import { Actions } from '@ngrx/effects';
+
+import { filter } from 'rxjs/operators';
+import { EntityAction, ofEntityOp, OP_ERROR, OP_SUCCESS } from 'ngrx-data';
 import { ToastService } from '../../core/toast.service';
 
 /** Report ngrx-data success/error actions as toast messages **/
 @Injectable()
 export class NgrxDataToastService {
-  constructor(actions$: EntityActions, toast: ToastService) {
+  constructor(actions$: Actions, toast: ToastService) {
     actions$
-      .where(ea => ea.op.endsWith(OP_SUCCESS) || ea.op.endsWith(OP_ERROR))
+      .pipe(
+        ofEntityOp(),
+        filter(
+          (ea: EntityAction) =>
+            ea.op.endsWith(OP_SUCCESS) || ea.op.endsWith(OP_ERROR)
+        )
+      )
       // this service never dies so no need to unsubscribe
       .subscribe(action =>
         toast.openSnackBar(`${action.entityName} action`, action.op)
